backend/requirements.txt
# Backend - FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# Async task queue
celery[redis]>=5.4.0
redis>=5.0.0

# GCP
google-cloud-bigquery>=3.20.0
google-cloud-bigquery-storage>=2.24.0
google-cloud-aiplatform>=1.60.0
google-auth>=2.29.0

# Data & tools
yfinance>=0.2.40
pandas>=2.2.0
requests>=2.31.0
httpx>=0.27.0

# PDF generation
fpdf2>=2.7.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Memory (BM25 lexical similarity for agent learning)
rank-bm25>=0.2.2
